Update Readme
Moved the exercise on adding a type signature from the 'Math' to
'Define Functions' section. Numbers are a special case in Elm so
it's confusing since the type is actually 'number' rather than Int
or Float.

Using the factorial function works since it uses List.range which
fixes the type to Int.